chore: update to new vault API

+++ programs/vault/src/certora/constants.rs
@@ -1,1 +1,1 @@
-pub const MAX_FEE_BPS: u16 = 10_000;
+pub const MAX_FEE_BPS: u64 = 10_000;
